+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 
+1.0.0
+-----
+* Stop excluding junit4 from testcontainers - check the README for the sordid details.
+* Pass host and port to the ConnectionInfo bean. We strongly recommend you prefer getDatasource or getUrl, these will be more portable in usage. We ran into
+a few use cases where this was handy, however.
+* LegacySingleInstancePostgresExtension to do the old Junit5 lifecycle behavior.
+* 60 seconds default startup wait.
+* Expose bind mounts (optional, none by default) in builder, currently hard coded as Read only.
+* Expose network (optional, none by default) in builder.
+
 1.0.0RC2
 -------
 * Restore Java 8 compatibility